Fix issue where the client crashes when goto app action is selected
This commit supplies the missing prop 'apps' to Table component
and EventSelector component so that they can properly display the
apps.

This commit also modifies the get-all-apps query logic such that
if the queried page is 0, it fetches all the apps.

It adds a filtering logic such that presently only the apps with
slugs are listed as targets for goto app action.