Update tabs.md
Formatting updates:

Replaced "Fx" and "FX" with "fx" 
For each h2 (created using ##), removed 24px padding-bottom. The div that wraps up the h2 headers and the section below the header should only have a "padding-top" of 24px and no "padding-bottom".

Content updates:
Replaced the word "widget" with "component" wherever applicable. 
Updated the title name from "Layout" to "Devices"
Updated the name of the first column from "Layout" to "Property"
Update the content in the "Description" and "Expected value" columns as per the given reference
Added a full stop at the end of the sentence in the "Description" column.
Added a divider before the "Styles" section reference
The changes are to be implemented in both the file